+++ src/views/Dashboard/components/WidgetContainer/WidgetContainer.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { Box } from "@mui/material";
 import { BoxProps } from "@mui/system";
 
-type IWidgetContainerProps = BoxProps;
-
 // TODO: handle better
 const drawerWidth = 240;
+
+type IWidgetContainerProps = BoxProps;
 
 // make slots for each added widget (slot1, slot2, etc and then each slot is a <Grid item>)?
 const WidgetContainer: React.FC<IWidgetContainerProps> = ({ children }) => {