+++ src/views/Login/components/OnboardingStepper/components/4.DisplayAddressStep/DisplayAddressStep.tsx
@@ -1,37 +1,36 @@
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useCallback, useEffect, useMemo, useState } from "react";
 import useAccount from "hooks/useAccount";
-import { Button, Checkbox, CheckboxProps, FormControlLabel, Typography } from "@mui/material";
+import { Button, Checkbox, CheckboxProps, FormControlLabel, Tooltip, Typography } from "@mui/material";
 import { IStepProps } from "../types";
 import { Alert } from "@mui/material";
 import { NavLink } from "react-router-dom";
 import RememberMeCheckbox from "views/Login/components/RememberMeCheckbox";
 import useLocalStorage from "hooks/useLocalStorage";
+import userEvent from "@testing-library/user-event";
 
 interface IUserNoticeGroupProps {
-  fetchUnderstandStatusFn: (userUnderstandArray: Array<boolean>) => void;
+  fetchUnderstandStatusFn: (userUnderstandArray: boolean) => void;
 }
 
 const UserNoticeGroup: React.FC<IUserNoticeGroupProps> = ({ fetchUnderstandStatusFn }) => {
   const [userUnderstandCheckboxStatus, setUserUnderstandCheckboxStatus] = useState<Array<boolean>>([false, false]);
 
-  const handleUserUnderandCheck = useCallback((newState: boolean, checkboxId: number) => {
-    let newArray: Array<boolean>;
-    //set user ack status for two different checkboxes
-    if (checkboxId === 0) {
+  const handleUserUnderandCheck = useCallback(
+    (newState: boolean, checkboxId: number) => {
       setUserUnderstandCheckboxStatus((prev) => {
-        newArray = [newState, prev[1]];
-        fetchUnderstandStatusFn(newArray);
-
+        console.log("handling new click", "newState:", newState, "checkboxId:", checkboxId, "prev:", prev);
+        let newArray: Array<boolean>;
+        if (checkboxId === 0) {
+          newArray = [newState, prev[1]];
+        } else {
+          newArray = [prev[0], newState];
+        }
+        setOverallUnderstandStatus(newArray, fetchUnderstandStatusFn);
         return newArray;
       });
-      return;
-    }
-    setUserUnderstandCheckboxStatus((prev) => {
-      newArray = [prev[0], newState];
-      fetchUnderstandStatusFn(newArray);
-      return newArray;
-    });
-  }, []);
+    },
+    [fetchUnderstandStatusFn]
+  );
 
   return (
     <>
@@ -54,7 +53,7 @@ const DisplayAccountStep: React.FC<IStepProps> = () => {
   const [userRememberState, setUserRememberState] = useState<boolean>(false);
   const [userUnderstandState, setUserUnderstandState] = useState<boolean>(false);
 
-  const [accounts, setAccounts] = useLocalStorage<Array<string>>("accounts", []); // stores user accounts in localStorage under "accounts" key
+  const [accounts, setAccounts] = useLocalStorage<Array<string>>("accounts", [] as Array<string>); // stores user accounts in localStorage under "accounts" key
 
   const handleLogin = useCallback(
     (e) => {
@@ -65,10 +64,19 @@ const DisplayAccountStep: React.FC<IStepProps> = () => {
         return;
       }
 
+      // user hasn't acked both confirm boxes yet
+      if (!userUnderstandState) {
+        e.preventDefault();
+        return;
+      }
+
+      console.log("checking if user wants to remember...", userRememberState);
       if (userRememberState) {
+        console.log("user wants to remember...");
         // address has not been previously saved
-        if (!accounts?.includes(accountRs)) {
+        if (!accounts.includes(accountRs)) {
           const newAccounts = [...accounts, accountRs];
+          console.log("setting new accounts:", newAccounts);
           setAccounts(newAccounts);
         }
       }
@@ -76,30 +84,31 @@ const DisplayAccountStep: React.FC<IStepProps> = () => {
       // flush the seed words from state as they are  no longer needed and could be a security risk if kept
       if (flushFn !== undefined) flushFn();
     },
-    [accountRs, accounts, setAccounts, userRememberState, flushFn]
+    [accountRs, accounts, setAccounts, userRememberState, flushFn, userUnderstandState]
   );
 
   const fetchUserRememberState = useCallback((isRememberedStatus: boolean) => {
+    console.log("setting remember status:", isRememberedStatus);
     setUserRememberState(isRememberedStatus);
   }, []);
 
-  // Debug
-  useMemo(() => {
-    console.log(userRememberState);
-  }, [userRememberState]);
-
-  // could I have a useMemo for (isUnderstanding) which is set from the UserNoticeGroup, so the UserNoticeGroup will
-  // perform the logic and return a single bool?
-  const fetchUserUnderstandState = useCallback((userUnderstandArray: Array<boolean>) => {
-    // all items are true
-    if (userUnderstandArray.every((element) => element === true)) {
-      console.log("all items true, user may proceed");
-      setUserUnderstandState(true);
-      return;
-    }
-    console.log("setting false");
-    setUserUnderstandState(false);
+  const fetchUserUnderstandState = useCallback((userUnderstands: boolean) => {
+    setUserUnderstandState(userUnderstands);
   }, []);
+
+  const UnderstandMemo = useMemo(() => {
+    if (userUnderstandState) {
+      return (
+        <NavLink to="/dashboard">
+          <Button size="large" onClick={(e) => handleLogin(e)} variant="contained">
+            Go To Dashboard
+          </Button>
+        </NavLink>
+      );
+    }
+
+    return <Typography>You must acknowledge the alert messages above before you can proceed</Typography>;
+  }, [handleLogin, userUnderstandState]);
 
   return (
     <>
@@ -108,13 +117,24 @@ const DisplayAccountStep: React.FC<IStepProps> = () => {
 
       <UserNoticeGroup fetchUnderstandStatusFn={fetchUserUnderstandState} />
       <FormControlLabel control={<RememberMeCheckbox fetchIsRememberedFn={fetchUserRememberState} />} label="Remember Account?" />
-      <NavLink to="/dashboard">
-        <Button size="large" onClick={(e) => handleLogin(e)} variant="contained" disabled={userRememberState}>
-          Go To Dashboard
-        </Button>
-      </NavLink>
+      {UnderstandMemo}
     </>
   );
 };
+
+//
+// Helper functions
+//
+
+// takes in an array of understand checkbox status' as well as a fetcher function that needs to know the overall status
+// and it returns a single bool
+function setOverallUnderstandStatus(statusArray: Array<boolean>, fetchFn: (status: boolean) => void) {
+  // all items are true
+  if (statusArray.every((element) => element === true)) {
+    fetchFn(true);
+    return;
+  }
+  fetchFn(false);
+}
 
 export default React.memo(DisplayAccountStep);